backend/requirements.txt
django~=4.1.0
django-cleanup
mysqlclient~=2.1
django-cors-headers~=3.11
djangorestframework~=3.13
django-filter~=22.1
djangorestframework_simplejwt~=5.1
django-q
django-redis
django_polymorphic

Pillow~=9.1
opencv-python-headless
scipy

django-labeltree

ammod-blob-detector
cvmodelz~=0.3


